Finished dependency list builder.
